src/main/java/com/sipios/refactoring/controller/ShoppingController.java
package com.sipios.refactoring.controller;

import com.sipios.refactoring.domain.Body;
import com.sipios.refactoring.service.ShoppingService;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.Date;

@RestController
@RequestMapping("/shopping")
public class ShoppingController {

    private final ShoppingService shoppingService;

    public ShoppingController(ShoppingService shoppingService) {
        this.shoppingService = shoppingService;
    }

    @PostMapping
    public String getPrice(@RequestBody Body body) {
        Date today = new Date();
        return shoppingService.calculatePrice(body, today);
    }
}
